refactor(@angular/cli): remove install-package under utils
Remove the lone `install-package.ts` file from tasks to utilities
